% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{auto_luria}
\alias{auto_luria}
\title{Run the luria pipeline}
\usage{
auto_luria(
  templateData,
  exclude.sheets = c("Example layout", "Column guide"),
  fill = 60,
  dilution = 1e+05,
  pool.as = NULL,
  exclude.reps = NULL,
  extract.mutrates = FALSE,
  export = TRUE,
  overwrite = FALSE
)
}
\arguments{
\item{templateData}{A luria Excel workbook template file, or a folder containing such.}

\item{exclude.sheets}{A character vector of sheet names (e.g. replicates) to skip. By default the
help sheets \emph{Example layout} and \emph{Column guide} is skipped, but more can be added.}

\item{fill}{A number indicating the true number of populations (i.e. wells) in each replicate.
Additional \emph{Selective} populations will be added with 0 CFU, up to this number. If \code{NULL}, no
filling will occur and only recorded populations will be used.
Defaults to \code{60}, the number of populations in a standard 96-well plate.}

\item{dilution}{A \emph{Count} population dilution factor to use for wrangling. The same factor will
be used for all replicates.
Defaults to \code{1e5}, the standard dilution rate for the \emph{A. baylyi} and \emph{E. coli} protocols.}

\item{pool.as}{What the pooled strain should be named. If \code{NULL} (the default), the strain is
named \emph{Combined}.}

\item{exclude.reps}{A character vector of replicates to skip.}

\item{extract.mutrates}{If \code{TRUE}, \code{extract_mutrates()} is also run. Defaults to \code{FALSE}.}

\item{export}{If \code{TRUE} (the default), outputs will be saved to a folder \verb{\\luria_output\\} in the
current directory.}

\item{overwrite}{If \code{TRUE}, output file(s) will be saved to the appropriate folder whether or not
existing file(s) are replaced.
Defaults to \code{FALSE}.}
}
\value{
A list object containing the outputs of the component functions. If a folder is supplied
to \code{templateData}, each raw file will produce a list of list objects corresponding to that raw
file.
}
\description{
Automatically process one or more template files to estimate mutation rates, export them, plot
the results, and (optionally) export the mutation rates.
}
\details{
A wrapper function that runs \code{\link[=wrangle_raw_data]{wrangle_raw_data()}}, \code{\link[=wrangle_clean_data]{wrangle_clean_data()}}, \code{\link[=run_fluxxer]{run_fluxxer()}},
and \code{\link[=plot_fluxxer]{plot_fluxxer()}} on a given template file or a folder containing multiple template files.
Argument values are inherited from those functions and can be specified in the same way.
\code{extract.mutrates} can optionally be set to \code{TRUE} to run \code{\link[=extract_mutrates]{extract_mutrates()}}, which will
return and/or export the mutation rate numbers for other uses.

If a folder of multiple template files is supplied to \code{templateData}, values to arguments such as
\code{exclude.sheets} are applied to all of them. To specify different argument values for multiple
template files, run this function on each template file separately with \code{overwrite = TRUE}.
}
\examples{
auto_luria(templateData, exclude.sheets = c("Example layout", "Column guide"),
           fill = 60, dilution = 1e5,
           pool.as = NULL, exclude.reps = NULL, extract.mutrates = FALSE,
           export = TRUE, overwrite = FALSE)

}
