% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{extract_mutrates}
\alias{extract_mutrates}
\title{Extract mutation rates in a reproducible manner}
\usage{
extract_mutrates(
  dataPath,
  export.method = "all",
  overwrite = FALSE,
  export.to = "."
)
}
\arguments{
\item{dataPath}{Path to a folder containing pooled mutation rates to extract.}

\item{export.method}{Takes value \code{"csv"}, which exports the data as a CSV, \code{"data"}, which
exports the data as an RData object; or \code{"all"}, which exports all formats.}

\item{overwrite}{If \code{TRUE}, output file(s) will be saved to the appropriate folder whether or not
existing file(s) are replaced.
Defaults to \code{FALSE}.}
}
\value{
A numeric vector with elements named by project.
}
\description{
Automatically get mutation rates from output data and save it.
}
\details{
Extracts pooled mutation rate estimates from multiple projects of pooled/unpooled file pairs
into a vector named by project, and exports them as a CSV of project name and estimates, and/or
an RData file. Likely only useful for extracting mutation rates from multiple projects at once.
This function does not take single files, and specifically looks for pooled estimate data as it
really only makes sense to use those.
}
\examples{
mutrates <- extract_mutrates(export.method = "all")

}
