% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{populate_rows}
\alias{populate_rows}
\title{Helper function for converting sheet data}
\usage{
populate_rows(
  currentData,
  currentCounts,
  currentMutants,
  countFraction,
  countPops
)
}
\arguments{
\item{currentData}{The current data to be populated and returned, usually \code{currentFrame}.}

\item{currentCounts}{\emph{Count} data from the current sheet.}

\item{currentMutants}{\emph{Selective} data from the current sheet.}

\item{countFraction}{\emph{fraction} value for the \emph{Count} populations; generated from \code{countFraction}
in the wrapper function.}

\item{countPops}{The number of \emph{Count} populations used.}
}
\value{
A dataframe of identical dimensions and headers as the input.
}
\description{
This function is called by \code{\link[=wrangle_old_raws]{wrangle_old_raws()}} inside the by-sheet loop. It takes a fluxxer-
formatted dataframe and uses the arguments to populate the rest of the rows, before returning
the dataframe.
}
\keyword{internal}
