% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{test_logticks}
\alias{test_logticks}
\title{Helper function to check for the annotation_logticks error}
\usage{
test_logticks(data)
}
\arguments{
\item{data}{The plotting data.}
}
\value{
A logical value indicating if logticks should be used.
}
\description{
This function is called by \code{\link[=wrangle_plot_data]{wrangle_plot_data()}}, and exists because because
\code{ggplot2::annotation_logticks} shits itself if the upper and lower CI differs by less than one
power of 10, but the error can't be easily caught once plotting begins. This function simply
checks of the upper and lower CI have a sufficient range, and returns a flag that is ultimately
passed to \code{\link[=plot_mutrates]{plot_mutrates()}} to enable (or disable) logtics.
}
\details{
If this doesn't work and it still throws an error, it might be because \code{annotation_logticks()}
still doesn't like if max(range) - min(range) >= 1 if they don't actually span two logticks
in the final plot. If that's the case, a more conservative range should be used, i.e.
floor(max(range)) - ceiling(min(range)), which collapses the difference to zero in this case.
}
\keyword{internal}
