% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{wrangle_old_raws}
\alias{wrangle_old_raws}
\title{(Legacy) Convert a fluctuation analysis Excel workbook}
\usage{
wrangle_old_raws(
  dataFile,
  countPops,
  countFract = c(P = 200, C = 200, D = 1e+05),
  exclude = "Summary",
  saveAs = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{dataFile}{The filename of the standard Excel workbook.}

\item{countPops}{An integer value indicating how many Count populations were used in each test.
This should be one value for all tests, and it should not account for failed Count plates.}

\item{countFract}{A named vector for calculating the fraction plated, F, of the Count
populations. Fraction is given as \eqn{F = P\div(CD)}, where P = volume plated, in uL;
C = volume of Count culture used for dilution, in uL; and D = dilution factor, expressed as an
integer (eg. \eqn{10^{3}}) and \strong{not} a ratio (eg. \eqn{10^{-3}}). The same vector is applied
to all tests. The default values are based on the original \emph{A. baylyi} protocol.}

\item{exclude}{A character vector of sheet names (i.e. replicates) to skip. By default the sheet
"Summary" is skipped, but more can be added.}

\item{saveAs}{A filename for saving the wrangled file.
Defaults to \code{NULL}, in which case the original filename will be used.}

\item{overwrite}{If \code{TRUE}, output file(s) will be saved to the appropriate folder whether or not
existing file(s) are replaced.
Defaults to \code{FALSE}.}
}
\description{
Wrangle a legacy-format standard fluctuation analysis Excel workbook, converting it for the
pipeline.
}
\details{
This function works on the old Excel workbook format to produce the standard CSV file with
columns \code{strain}, \code{plate}, \code{fraction}, and \code{CFU}. Templates obtained by calling
\code{\link[=get_template]{get_template()}} must be wrangled with \code{\link[=wrangle_raw_data]{wrangle_raw_data()}} instead. The next function in
the pipeline is \code{\link[=wrangle_clean_data]{wrangle_clean_data()}}.
}
\examples{
wrangle_old_raws(dataFile = "./data/raws/FLUCTEST 1 2020 09 24.xlsx",
                 countPops = 4,
                 exclude = c("Summary", "Rep 0", "Rep 13"),
                 saveAs = "RIF_Aug2023")

}
