% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{calculate_mut_rate}
\alias{calculate_mut_rate}
\title{Calculate mutation rates.}
\usage{
calculate_mut_rate(
  filename = NULL,
  outputPath,
  outputPrefix = "",
  comparisons = FALSE
)
}
\arguments{
\item{filename}{A correctly-formatted CSV to estimate mutation rates from.}

\item{outputPath}{The directory to save output files to.}

\item{outputPrefix}{A prefix to use when saving the output file, which will be named like
\verb{<outputPrefix>.output.csv}}

\item{comparisons}{Logical. Indicates if a comparison file should be generated.}
}
\description{
A minimally-modified version of \code{calculateMutRate()} from the original
\href{https://github.com/barricklab/barricklab/blob/master/fluxxer.R}{\code{fluxxer.R}} script.
}
\details{
This function uses the rSalvador package to obtain the maximum likelihood estimates for
mutation rates based on the Luria-Delbruck distribution.
Do not call this function manually - use \code{\link[=run_fluxxer]{run_fluxxer()}} instead, which supplies the
appropriate arguments. This version does not produce plots, as it is handled by the pipeline
function \code{\link[=plot_fluxxer]{plot_fluxxer()}}.
}
