% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{extract_mutrates}
\alias{extract_mutrates}
\title{Extract mutation rates in a reproducible manner.}
\usage{
extract_mutrates(method = "save", overwrite = FALSE)
}
\arguments{
\item{method}{Takes value \code{"save"}, which exports the data as an RData object; \code{"return"}, which
returns the mutation rate as an RData object; or \code{"both"}, which does both.}

\item{overwrite}{If \code{TRUE}, output file(s) will be saved to the appropriate folder whether or not
existing file(s) are replaced.
Defaults to \code{FALSE}.}
}
\value{
If \code{method} is \code{"return"} or \code{"both"}, a numeric vector with elements named by project.
}
\description{
Automatically get mutation rates from output data and save it as an RData object.
}
\details{
This function was originally designed to quickly export pooled mutation rate estimates for
simulation work, making it easier to track estimate versions and removing the need for copy +
pasting. It is likely only useful if you have multiple projects to extract numbers from.
This function does not take single files, as it really only makes sense to take the pooled
estimate anyway.
}
\keyword{internal}
