% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{get_template}
\alias{get_template}
\title{Save a copy of the fluctuation analysis data template.}
\usage{
get_template(saveTo = ".", saveAs = "template", version = "latest")
}
\arguments{
\item{saveTo}{Path to save the template file. Defaults to the project root.}

\item{saveAs}{Filename to use. The default filename is "template.xlsx".}

\item{version}{Which version of the template to use. Defaults to \code{latest}; \code{old} saves a copy of
the old template - use this at your own risk.}
}
\description{
Use this function to save a copy of the fluctuation analysis Excel data template.
}
\examples{
get_template("./data", "my_data_template")

}
