% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{plot_mutrates}
\alias{plot_mutrates}
\title{The core plotting function.}
\usage{
plot_mutrates(data, levelOrder, log = NULL)
}
\arguments{
\item{data}{A wrangled data object to use for plotting.}

\item{levelOrder}{A string vector defining the order that replicates will appear on the x-axis.}

\item{log}{A logical value indicating if the data will trigger the \code{\link{annotation_logticks}} error.
Defaults to \code{NULL}; logticks are applied if \code{TRUE}.}
}
\value{
A ggplot2 object.
}
\description{
Produce a ggplot object with or without a log-scale y-axis.
}
\details{
This function does nothing but produce the \emph{luria} plot. Factoring the replicates is done by
\code{\link[=refactor_reps]{refactor_reps()}}, and switching the log y-axis is done through a logical value supplied by
\code{\link[=test_logticks]{test_logticks()}}. \code{\link[=plot_fluxxer]{plot_fluxxer()}} calls this function every time a plot is generated.
}
\keyword{internal}
