% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{prep_export}
\alias{prep_export}
\title{Helper function for setting up and checking folder structure and pathing.}
\usage{
prep_export(mode = NULL, overwrite = FALSE)
}
\arguments{
\item{mode}{Defined by the calling function to determine what output folder is appropriate and
how to handle it. Takes values \code{"wrangled"}, \code{"analyzed"}, \code{"plots"}, or \code{"mutrates"}.}

\item{overwrite}{If \code{TRUE}, output file(s) will be saved to the appropriate folder whether or not
existing file(s) are replaced.
Defaults to \code{FALSE}.}
}
\value{
A string indicating the output folder path.
}
\description{
This function is called by all frontend functions that produce file outputs, namely
\code{\link[=wrangle_raw_data]{wrangle_raw_data()}}, \code{\link[=wrangle_clean_data]{wrangle_clean_data()}}, \code{\link[=run_fluxxer]{run_fluxxer()}}, and \code{\link[=plot_fluxxer]{plot_fluxxer()}}.
It always attempts to create the top-level output folder \verb{./output/} whenever it's called,
and enforces output folder structure by (a) checking if the output subfolder exists and if
its contents can be overwritten, and (b) returning the output path for use by the calling
function so it only has to be defined once. \code{overwrite} is determined by the user in the
frontend functions; if \code{FALSE}, files cannot be written to the output folder if it contains
valid files.
}
\keyword{internal}
