% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{refactor_reps}
\alias{refactor_reps}
\title{Helper function for reordering replicate names.}
\usage{
refactor_reps(data, pooledPrefix = NULL)
}
\arguments{
\item{data}{The plotting data.}

\item{pooledPrefix}{The pooled replicate name.}
}
\value{
A string vector of the replicate names in the correct order, with \code{pooledPrefix} first.
}
\description{
This function is called by \code{\link[=wrangle_plot_data]{wrangle_plot_data()}}. It assumes that each unpooled replicate is
named in the format \verb{<rep><number>}, where \verb{<rep>} can be any string, with or without spaces, as
long as it's identical for all replicates. Because it takes the pooled replicate prefix
explicitly via \code{pooledPrefix}, theoretically any number of nonconforming rep names can be used
as long as they are all explicitly defined in a vector - although there should really only be
one pooled rep.
}
\keyword{internal}
