% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{run_fluxxer}
\alias{run_fluxxer}
\title{Analyze the data and estimate mutation rates.}
\usage{
run_fluxxer(file = NULL, comparisons = TRUE, overwrite = FALSE)
}
\arguments{
\item{file}{The filename of a correctly-formatted CSV file to estimate mutation rates from.
Supplying this will cause the function to run in single-file mode.}

\item{comparisons}{Logical value indicating if a comparison file should be generated. If \code{file}
is also supplied, the function will generate a comparison even if only one strain is available.}

\item{overwrite}{If \code{TRUE}, output file(s) will be saved to the appropriate folder whether or not
existing file(s) are replaced.
Defaults to \code{FALSE}.}
}
\description{
Estimate mutation rates from wrangled fluctuation analysis data, using an individual file or
as the next step in the pipeline, and output them as CSV files.
}
\details{
This function calls \code{\link[=calculate_mut_rate]{calculate_mut_rate()}} to estimate mutation rates. It can run on a single
file, or cycle automatically through all files saved to \verb{./output/wrangled/} by the previous step
in the pipeline. For every input file, a matching output file ending in \code{.output.csv} will be
saved to \code{./output/analyzed}. If \code{comparisons} is \code{TRUE}, an additional file will be generated
ending in \code{.comparisons.csv}; this file contains pairwise comparisons between all replicates and
p-values indicating how different the mutation rates are from each other. Supplying the \code{file}
argument runs this function in single-file mode. In standard pipeline mode, comparisons are only
made using unpooled data.
}
\examples{
run_fluxxer(comparisons = TRUE)

}
