% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{wrangle_clean_data}
\alias{wrangle_clean_data}
\title{Create standardized pooled and unpooled data files for the pipeline.}
\usage{
wrangle_clean_data(
  dataFile = NULL,
  poolAs = NULL,
  exclude = NULL,
  saveAs = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{dataFile}{A CSV file with the standard headers \code{strain}, \code{plate}, \code{fraction}, and \code{CFU}.}

\item{poolAs}{What the pooled strain should be named.
Defaults to \code{NULL}, in which case the strain is just named \emph{Combined}.}

\item{exclude}{A character vector of replicates to skip.}

\item{saveAs}{A filename for saving the wrangled file.
Defaults to \code{NULL}, in which case the original filename will be used.}

\item{overwrite}{If \code{TRUE}, output file(s) will be saved to the appropriate folder whether or not
existing file(s) are replaced.
Defaults to \code{FALSE}.}
}
\value{
Two tidy CSVs with columns \code{strain}, \code{plate}, \code{fraction}, and \code{CFU}.
}
\description{
Prepare CSV files previously wrangled by \code{\link[=wrangle_raw_data]{wrangle_raw_data()}} for the rest of the
pipeline. Correctly-formatted CSVs otherwise created (e.g. if the Excel template was not
used) can also be prepared for the pipeline with this function.
}
\details{
This function produces two CSV files, one with all replicates pooled into one strain, and another
CSV with the replicates kept separate. Both CSV files are tidy tables in a standardized format,
which is required by \code{\link[=run_fluxxer]{run_fluxxer()}}, the next function in the pipeline.
}
\section{Output CSV format}{

The resulting CSV is a tidy table with four columns:
\itemize{
\item \code{strain} - The name of the replicate
\item \code{plate} - The type of plate, either \emph{Count} or \emph{Selective}
\item \code{fraction} - The fraction of culture plated (see below)
\item \code{CFU} - The observed number of colonies on the plate.
}
Additionally, fraction is given as \eqn{F = P\div(CD)}, where P = volume plated, in uL;
C = volume of Count culture used for dilution, in uL; and D = dilution factor. Note that D is
expressed as an integer (eg. \eqn{10^{3}}) and not a ratio (eg. \eqn{10^{-3}}). Fraction is
calculated individually for every observation in the raw file.
}

\examples{
wrangle_clean_data(dataFile = "./data/raws/FLUCTEST 1 2020 09 24.csv",
                   poolAs = "AB3",
                   exclude = c("Rep 0", "Rep 13"),
                   saveAs = "RIF_Aug2023")
}
