% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{wrangle_plot_data}
\alias{wrangle_plot_data}
\title{Wrangle fluxxer outputs for plotting.}
\usage{
wrangle_plot_data(file = NULL, projectName = NULL, inputPath = NULL)
}
\arguments{
\item{file}{The filename of the analyzed .output.csv to wrangle. If supplied, the function will
run in single file mode.}

\item{projectName}{A project prefix as understood by the pipeline. If supplied, the function will
run in standard pipeline mode.}

\item{inputPath}{The path for the pipeline output folder from which files are automatically
retrieved. This must be supplied by \code{\link[=plot_fluxxer]{plot_fluxxer()}} if \code{projectName} is supplied.}
}
\value{
A list with: \code{data}, the plotting data; \code{levels}, a vector of replicate names in the
correct order; and \code{log}, which if \code{TRUE} will signal \code{\link[=plot_mutrates]{plot_mutrates()}} to use logticks.
}
\description{
Internal function called by \code{\link[=plot_fluxxer]{plot_fluxxer()}}. It imports and wrangles analyzed data to prepare
it for plotting.
}
\details{
This is a single/standard mode function, taking either an explicit fluxxer .output.csv file, or
automatically retrieving these files from /output/analyzed/. It (a) produces a data object of
combined pooled and unpooled data; (b) calls \code{\link[=refactor_reps]{refactor_reps()}} to rearrange replicate names for
plotting; and (c) calls \code{\link[=test_logticks]{test_logticks()}} to detect if the data will cause
\code{\link[ggplot2:annotation_logticks]{ggplot2::annotation_logticks}} to throw an error due to error bars not spanning 2 logticks.
}
\examples{
wrangle_plot_data(file = "./data/old_fluc_stuff/FLUCTEST1.output.csv",
                  projectName = NULL,
                  inputPath = NULL)

}
\keyword{internal}
