% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{wrangle_raw_data}
\alias{wrangle_raw_data}
\title{Convert a fluctuation analysis Excel workbook.}
\usage{
wrangle_raw_data(
  templateFile,
  exclude = c("Example layout", "Column guide"),
  fill = 60,
  dilution = 1e+05,
  saveAs = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{templateFile}{The filename of the standard Excel workbook.}

\item{exclude}{A character vector of sheet names (e.g. replicates) to skip. By default the
help sheets \emph{Example layout} and \emph{Column guide} is skipped, but more can be added.}

\item{fill}{A number indicating the true number of populations (i.e. wells) in each replicate.
Additional \emph{Selective} populations will be added with 0 CFU, up to this number. If \code{NULL}, no
filling will occur and only recorded populations will be used.
Defaults to \code{60}, the number of populations in a standard 96-well plate.}

\item{dilution}{A \emph{Count} population dilution factor to use for wrangling. The same factor will
be used for all replicates.
Defaults to \code{1e5}, the standard dilution rate for the \emph{A. baylyi} and \emph{E. coli} protocols.}

\item{saveAs}{A filename for saving the wrangled file.
Defaults to \code{NULL}, in which case the original filename will be used.}

\item{overwrite}{If \code{TRUE}, output file(s) will be saved to the appropriate folder whether or not
existing file(s) are replaced.
Defaults to \code{FALSE}.}
}
\description{
Wrangle a standard fluctuation analysis Excel workbook, converting it for the pipeline.
}
\details{
This function converts raw data contained in an Excel workbook template into a standardized CSV
file which is required by downstream functions. The next function in the pipeline is
\code{\link[=wrangle_clean_data]{wrangle_clean_data()}}.
}
\section{Output CSV format}{

The resulting CSV is a tidy table with four columns:
\itemize{
\item \code{strain} - The name of the replicate
\item \code{plate} - The type of plate, either \emph{Count} or \emph{Selective}
\item \code{fraction} - The fraction of culture plated (see below)
\item \code{CFU} - The observed number of colonies on the plate.
}
Additionally, fraction is given as \eqn{F = P\div(CD)}, where P = volume plated, in uL;
C = volume of Count culture used for dilution, in uL; and D = dilution factor. Note that D is
expressed as an integer (eg. \eqn{10^{3}}) and not a ratio (eg. \eqn{10^{-3}}). Fraction is
calculated individually for every observation in the raw file.
}

\examples{
wrangle_raw_data(templateFile = "./data/raws/FLUCTEST 1 2020 09 24.xlsx",
                 exclude = c("Rep 0", "Rep 13"),
                 saveAs = "RIF_Aug2023")

}
